Create getHumanChoice() function
Function will prompt user to type "rock", "paper", or "scissors". If
user clicks "cancel", the user will be prompted with the question to end
the game. If they click ok, the function is returned, otherwise they
will return to the original prompt. The function then confirms a valid
input and then returns that input.